+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * :fire: The parameters of the Bundle type of the Activityx start(\*, Class, Bundle) method are now used as pass parameters
 * :sparkles: Activityx adds safeStart(\*, Class, Bundle) and safeStart(\*, Class) method
 * :white_check_mark: Improve Activityx testing
+* :fire: ResizeCalculator rename to Resizex
 
 
 ## v0.5.1